# Remove ProtocolReader async state machine allocations from sync path