This improves #20 by a decent amount when the data is available from the pipe synchronously. Benchmarks look pretty good here:

Before:

```
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
AMD Ryzen Threadripper 1950X, 1 CPU, 32 logical and 16 physical cores
.NET Core SDK=3.1.100
  [Host]     : .NET Core 3.1.0 (CoreCLR 4.700.19.56402, CoreFX 4.700.19.56404), X64 RyuJIT
  DefaultJob : .NET Core 3.1.0 (CoreCLR 4.700.19.56402, CoreFX 4.700.19.56404), X64 RyuJIT


|                                  Method | MessageSize |     Mean |   Error |  StdDev | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------------------- |------------ |---------:|--------:|--------:|------:|--------:|------:|------:|------:|----------:|
|                                PipeOnly |          10 | 343.8 ns | 2.71 ns | 2.40 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   ReadProtocolWithWholeMessageAvailable |          10 | 404.4 ns | 8.00 ns | 9.21 ns |  1.18 |    0.03 |     - |     - |     - |         - |
| ReadProtocolWithPartialMessageAvailable |          10 | 549.2 ns | 1.46 ns | 1.36 ns |  1.60 |    0.01 |     - |     - |     - |         - |
|                                         |             |          |         |         |       |         |       |       |       |           |
|                                PipeOnly |         100 | 343.6 ns | 1.73 ns | 1.45 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   ReadProtocolWithWholeMessageAvailable |         100 | 409.3 ns | 3.05 ns | 2.70 ns |  1.19 |    0.01 |     - |     - |     - |         - |
| ReadProtocolWithPartialMessageAvailable |         100 | 535.6 ns | 3.92 ns | 3.27 ns |  1.56 |    0.01 |     - |     - |     - |         - |
|                                         |             |          |         |         |       |         |       |       |       |           |
|                                PipeOnly |        1000 | 358.6 ns | 3.24 ns | 2.70 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   ReadProtocolWithWholeMessageAvailable |        1000 | 491.5 ns | 2.64 ns | 2.06 ns |  1.37 |    0.01 |     - |     - |     - |         - |
| ReadProtocolWithPartialMessageAvailable |        1000 | 642.1 ns | 7.15 ns | 6.69 ns |  1.79 |    0.02 |     - |     - |     - |         - |
```

After:

```
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
AMD Ryzen Threadripper 1950X, 1 CPU, 32 logical and 16 physical cores
.NET Core SDK=3.1.100
  [Host]     : .NET Core 3.1.0 (CoreCLR 4.700.19.56402, CoreFX 4.700.19.56404), X64 RyuJIT
  DefaultJob : .NET Core 3.1.0 (CoreCLR 4.700.19.56402, CoreFX 4.700.19.56404), X64 RyuJIT


|                                  Method | MessageSize |     Mean |    Error |   StdDev | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------------------- |------------ |---------:|---------:|---------:|------:|--------:|------:|------:|------:|----------:|
|                                PipeOnly |          10 | 368.6 ns |  4.80 ns |  4.49 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   ReadProtocolWithWholeMessageAvailable |          10 | 379.6 ns |  3.07 ns |  2.87 ns |  1.03 |    0.02 |     - |     - |     - |         - |
| ReadProtocolWithPartialMessageAvailable |          10 | 501.8 ns |  8.37 ns |  7.83 ns |  1.36 |    0.02 |     - |     - |     - |         - |
|                                         |             |          |          |          |       |         |       |       |       |           |
|                                PipeOnly |         100 | 365.3 ns |  2.26 ns |  2.12 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   ReadProtocolWithWholeMessageAvailable |         100 | 382.9 ns |  6.48 ns |  6.06 ns |  1.05 |    0.02 |     - |     - |     - |         - |
| ReadProtocolWithPartialMessageAvailable |         100 | 511.0 ns | 10.06 ns | 10.33 ns |  1.40 |    0.03 |     - |     - |     - |         - |
|                                         |             |          |          |          |       |         |       |       |       |           |
|                                PipeOnly |        1000 | 385.3 ns |  6.52 ns |  5.45 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|   ReadProtocolWithWholeMessageAvailable |        1000 | 467.8 ns |  7.81 ns |  7.31 ns |  1.21 |    0.03 |     - |     - |     - |         - |
| ReadProtocolWithPartialMessageAvailable |        1000 | 597.6 ns |  6.44 ns |  6.02 ns |  1.55 |    0.03 |     - |     - |     - |         - |
```